Prefab Positioning and Scaling is done!
I wrote another script and I named it as ScaleTest. In this script I tried to change scale of an object to learn how it is possible. Later on, I went to object pooling script and applied what I learn in there. I managed to change the prefab's scale and positions. (ScaleTest script is not used in anywhere right now, it can be deleted or remained still)